Add plug search to SocketDetails
