+++ src/app/item-popup/SocketDetails.m.scss
@@ -15,6 +15,7 @@
     display: flex;
     flex-direction: row;
     align-items: center;
+    min-height: 24px;
     > * {
       margin-right: 8px;
       line-height: 0;
