+++ src/app/loadout/plug-drawer/PlugDrawer.tsx
@@ -1,7 +1,7 @@
 import { languageSelector } from 'app/dim-api/selectors';
 import { PluggableInventoryItemDefinition } from 'app/inventory/item-types';
 import { useD2Definitions } from 'app/manifest/selectors';
-import { startWordRegexp } from 'app/search/search-filters/freeform';
+import { createPlugSearchPredicate } from 'app/search/plug-search';
 import { SearchInput } from 'app/search/SearchInput';
 import { useIsPhonePortrait } from 'app/shell/selectors';
 import { Comparator, compareBy } from 'app/utils/comparators';
@@ -147,26 +147,13 @@ export default function PlugDrawer({
       return plugSets;
     }
 
-    const regexp = startWordRegexp(query, language);
-    const searchFilter = (plug: PluggableInventoryItemDefinition) =>
-      regexp.test(plug.displayProperties.name) ||
-      regexp.test(plug.displayProperties.description) ||
-      regexp.test(plug.itemTypeDisplayName) ||
-      plug.perks.some((perk) => {
-        const perkDef = defs.SandboxPerk.get(perk.perkHash);
-        return (
-          perkDef &&
-          (regexp.test(perkDef.displayProperties.name) ||
-            regexp.test(perkDef.displayProperties.description) ||
-            regexp.test(perk.requirementDisplayString))
-        );
-      });
+    const searchFilter = createPlugSearchPredicate(query, language, defs);
 
     return plugSets.map((plugSet) => ({
       ...plugSet,
       plugs: plugSet.plugs.filter(searchFilter),
     }));
-  }, [query, plugSets, defs.SandboxPerk, language]);
+  }, [query, plugSets, defs, language]);
 
   if (sortPlugGroups) {
     queryFilteredPlugSets.sort(sortPlugGroups);
